feat(lab-eixf.1): Aurora foundations cleanup — shared tokens module + stale globals deletion
- Create components/aurora/tokens.ts with corrected typography constants:
  AURORA_DISPLAY_1 (34px/-0.045em), AURORA_DISPLAY_2 (19px/-0.02em),
  AURORA_DISPLAY_NUMBER, AURORA_MUTED_LABEL (font-bold per contract),
  AURORA_PAGE_FRAME, AURORA_PAGE_SHELL, AURORA_MEDIUM_PANEL,
  AURORA_STRONG_PANEL, AURORA_CONTROL_SURFACE, AURORA_MESSAGE_SURFACE,
  pillTone() and controlTone() with hex removed (now use Aurora tokens)
- Add components/aurora/README.md with one-way dependency rule
- Slim log-theme.ts to logs-only: AURORA_LEVEL_TEXT + AURORA_TAIL_ROW
- Update gateway-theme.ts: remove log-theme imports and re-exports,
  import AURORA_CONTROL_SURFACE/controlTone from aurora/tokens
- Migrate all 11 log-theme call sites and gateway component call sites
  to import shared constants from @/components/aurora/tokens
- Delete stale apps/gateway-admin/styles/globals.css (not imported)
- Also includes eixf.2: .dark shadcn token remap + --aurora-text-muted
  lightened (#90a9b9 → #a7bcc9, WCAG AA), shadow/highlight @theme
  forwarding, --aurora-success added to :root
