+++ apps/gateway-admin/components/aurora/README.md
@@ -1,0 +1,3 @@
+Shared Aurora design-system constants. Source of truth for token naming is `app/globals.css` (CSS vars) and `docs/design-system-contract.md` (semantics).
+
+**One-way dependency rule:** `components/aurora/**` must NOT import from `components/ui/**`. Token module depends on nothing in the primitive tree. Primitives may depend on this module.
